add ng package module
